requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
chromadb
pypdf
python-docx
pandas
openpyxl
pillow
pytesseract
google-generativeai
python-dotenv
